Split API/Providers concerns in VPN modules
